@@ -4,13 +4,16 @@
   "private": true,
   "dependencies": {
     "@testing-library/jest-dom": "^5.16.2",
-    "@testing-library/react": "^12.1.3",
+    "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
-    "bootstrap": "^5.1.3",
-    "react": "^17.0.2",
-    "react-bootstrap": "^2.2.0",
-    "react-dom": "^17.0.2",
+    "bootstrap": "^5.3.6",
+    "framer-motion": "^7.10.3",
+    "lucide-react": "^0.510.0",
+    "react": "^18.3.1",
+    "react-bootstrap": "^2.10.10",
+    "react-dom": "^18.3.1",
     "react-icons": "^4.3.1",
+    "react-intersection-observer": "^9.16.0",
     "react-owl-carousel": "^2.3.3",
     "react-router-dom": "^6.2.2",
     "react-scripts": "5.0.0",
